--- planit-react/src/pages/CreateEvent.jsx
@@ -1,14 +1,0 @@
-import Button from "../components/button/Button.jsx";
-import NavBar from "../components/navbar/Navbar.jsx";
-
-const CreateEvent = () => {
-    return (
-      <div>
-      <NavBar></NavBar>
-     <p>I AM CREATE EVENT PAGE</p>
-     <Button className={"special"}></Button>
-     </div>
-    );
-  };
-  
-  export default CreateEvent;
