+++ planit-react/src/pages/CreateEvent/CreateEvent.jsx
@@ -1,0 +1,14 @@
+import Button from "../../components/button/Button.jsx";
+import NavBar from "../../components/navbar/NavBar.jsx";
+
+const CreateEvent = () => {
+  return (
+    <div>
+      <NavBar></NavBar>
+      <p>I AM CREATE EVENT PAGE</p>
+      <Button className={"special"}></Button>
+    </div>
+  );
+};
+
+export default CreateEvent;
